refactor: Name 클래스 추가 및 관련 로직 변경
