+++ src/main/java/carracing/view/ResultView.java
@@ -31,7 +31,7 @@ public class ResultView {
     }
 
     public void print(Car car) {
-        System.out.printf(CAR_NAME, car.getName());
+        System.out.printf(CAR_NAME, car.getName().getAlias());
         for (int i = 0; i < car.getDrivingHistory(); i++) {
             System.out.print(MOVED);
         }
